move tests back to the toplevel recovery module
