+++ quiche/src/recovery/congestion/recovery.rs
@@ -61,7 +61,7 @@ use crate::recovery::MAX_PACKET_THRESHOLD;
 use crate::recovery::MAX_PTO_PROBES_COUNT;
 
 #[derive(Default)]
-struct RecoveryEpoch {
+pub struct RecoveryEpoch {
     /// The time the most recent ack-eliciting packet was sent.
     time_of_last_ack_eliciting_packet: Option<Instant>,
 
@@ -75,10 +75,10 @@ struct RecoveryEpoch {
 
     /// An association of packet numbers in a packet number space to information
     /// about them.
-    sent_packets: VecDeque<Sent>,
+    pub sent_packets: VecDeque<Sent>,
 
-    loss_probes: usize,
-    in_flight_count: usize,
+    pub loss_probes: usize,
+    pub in_flight_count: usize,
 
     acked_frames: Vec<frame::Frame>,
     lost_frames: Vec<frame::Frame>,
@@ -296,17 +296,17 @@ impl RecoveryEpoch {
 }
 
 pub struct Recovery {
-    epochs: [RecoveryEpoch; packet::Epoch::count()],
+    pub epochs: [RecoveryEpoch; packet::Epoch::count()],
 
     loss_timer: LossDetectionTimer,
 
-    pto_count: u32,
+    pub pto_count: u32,
 
-    rtt_stats: RttStats,
+    pub rtt_stats: RttStats,
 
     pub lost_spurious_count: usize,
 
-    pkt_thresh: u64,
+    pub pkt_thresh: u64,
 
     time_thresh: f64,
 
@@ -808,7 +808,7 @@ impl Recovery {
         }
     }
 
-    fn detect_lost_packets(
+    pub fn detect_lost_packets(
         &mut self, epoch: packet::Epoch, now: Instant, trace_id: &str,
     ) -> (usize, usize) {
         let loss_delay = cmp::max(self.rtt_stats.latest_rtt, self.rtt())
@@ -953,964 +953,4 @@ pub struct Acked {
     pub first_sent_time: Instant,
 
     pub is_app_limited: bool,
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::recovery::congestion::PACING_MULTIPLIER;
-    use crate::CongestionControlAlgorithm;
-    use smallvec::smallvec;
-    use std::str::FromStr;
-
-    #[test]
-    fn lookup_cc_algo_ok() {
-        let algo = CongestionControlAlgorithm::from_str("reno").unwrap();
-        assert_eq!(algo, CongestionControlAlgorithm::Reno);
-    }
-
-    #[test]
-    fn lookup_cc_algo_bad() {
-        assert_eq!(
-            CongestionControlAlgorithm::from_str("???"),
-            Err(crate::Error::CongestionControl)
-        );
-    }
-
-    #[test]
-    fn loss_on_pto() {
-        let mut cfg = crate::Config::new(crate::PROTOCOL_VERSION).unwrap();
-        cfg.set_cc_algorithm(CongestionControlAlgorithm::Reno);
-
-        let mut r = Recovery::new(&cfg);
-
-        let mut now = Instant::now();
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-
-        // Start by sending a few packets.
-        let p = Sent {
-            pkt_num: 0,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 1);
-        assert_eq!(r.bytes_in_flight, 1000);
-
-        let p = Sent {
-            pkt_num: 1,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 2000);
-
-        let p = Sent {
-            pkt_num: 2,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 3);
-        assert_eq!(r.bytes_in_flight, 3000);
-
-        let p = Sent {
-            pkt_num: 3,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 4);
-        assert_eq!(r.bytes_in_flight, 4000);
-
-        // Wait for 10ms.
-        now += Duration::from_millis(10);
-
-        // Only the first 2 packets are acked.
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(0..2);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                25,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((0, 0, 2 * 1000))
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 2000);
-        assert_eq!(r.congestion.lost_count, 0);
-
-        // Wait until loss detection timer expires.
-        now = r.loss_detection_timer().unwrap();
-
-        // PTO.
-        r.on_loss_detection_timeout(HandshakeStatus::default(), now, "");
-        assert_eq!(r.epochs[packet::Epoch::Application].loss_probes, 1);
-        assert_eq!(r.congestion.lost_count, 0);
-        assert_eq!(r.pto_count, 1);
-
-        let p = Sent {
-            pkt_num: 4,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 3);
-        assert_eq!(r.bytes_in_flight, 3000);
-
-        let p = Sent {
-            pkt_num: 5,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 4);
-        assert_eq!(r.bytes_in_flight, 4000);
-        assert_eq!(r.congestion.lost_count, 0);
-
-        // Wait for 10ms.
-        now += Duration::from_millis(10);
-
-        // PTO packets are acked.
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(4..6);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                25,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((2, 2000, 2 * 1000))
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 4);
-        assert_eq!(r.bytes_in_flight, 0);
-
-        assert_eq!(r.congestion.lost_count, 2);
-
-        // Wait 1 RTT.
-        now += r.rtt();
-
-        r.detect_lost_packets(packet::Epoch::Application, now, "");
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-    }
-
-    #[test]
-    fn loss_on_timer() {
-        let mut cfg = crate::Config::new(crate::PROTOCOL_VERSION).unwrap();
-        cfg.set_cc_algorithm(CongestionControlAlgorithm::Reno);
-
-        let mut r = Recovery::new(&cfg);
-
-        let mut now = Instant::now();
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-
-        // Start by sending a few packets.
-        let p = Sent {
-            pkt_num: 0,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 1);
-        assert_eq!(r.bytes_in_flight, 1000);
-
-        let p = Sent {
-            pkt_num: 1,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 2000);
-
-        let p = Sent {
-            pkt_num: 2,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 3);
-        assert_eq!(r.bytes_in_flight, 3000);
-
-        let p = Sent {
-            pkt_num: 3,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 4);
-        assert_eq!(r.bytes_in_flight, 4000);
-
-        // Wait for 10ms.
-        now += Duration::from_millis(10);
-
-        // Only the first 2 packets and the last one are acked.
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(0..2);
-        acked.insert(3..4);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                25,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((0, 0, 3 * 1000))
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 1000);
-        assert_eq!(r.congestion.lost_count, 0);
-
-        // Wait until loss detection timer expires.
-        now = r.loss_detection_timer().unwrap();
-
-        // Packet is declared lost.
-        r.on_loss_detection_timeout(HandshakeStatus::default(), now, "");
-        assert_eq!(r.epochs[packet::Epoch::Application].loss_probes, 0);
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 0);
-
-        assert_eq!(r.congestion.lost_count, 1);
-
-        // Wait 1 RTT.
-        now += r.rtt();
-
-        r.detect_lost_packets(packet::Epoch::Application, now, "");
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-    }
-
-    #[test]
-    fn loss_on_reordering() {
-        let mut cfg = crate::Config::new(crate::PROTOCOL_VERSION).unwrap();
-        cfg.set_cc_algorithm(CongestionControlAlgorithm::Reno);
-
-        let mut r = Recovery::new(&cfg);
-
-        let mut now = Instant::now();
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-
-        // Start by sending a few packets.
-        let p = Sent {
-            pkt_num: 0,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 1);
-        assert_eq!(r.bytes_in_flight, 1000);
-
-        let p = Sent {
-            pkt_num: 1,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 2000);
-
-        let p = Sent {
-            pkt_num: 2,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 3);
-        assert_eq!(r.bytes_in_flight, 3000);
-
-        let p = Sent {
-            pkt_num: 3,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 4);
-        assert_eq!(r.bytes_in_flight, 4000);
-
-        // Wait for 10ms.
-        now += Duration::from_millis(10);
-
-        // ACKs are reordered.
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(2..4);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                25,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((1, 1000, 1000 * 2))
-        );
-
-        now += Duration::from_millis(10);
-
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(0..2);
-
-        assert_eq!(r.pkt_thresh, INITIAL_PACKET_THRESHOLD);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                25,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((0, 0, 1000))
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-        assert_eq!(r.bytes_in_flight, 0);
-
-        // Spurious loss.
-        assert_eq!(r.congestion.lost_count, 1);
-        assert_eq!(r.lost_spurious_count, 1);
-
-        // Packet threshold was increased.
-        assert_eq!(r.pkt_thresh, 4);
-
-        // Wait 1 RTT.
-        now += r.rtt();
-
-        r.detect_lost_packets(packet::Epoch::Application, now, "");
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-    }
-
-    #[test]
-    fn pacing() {
-        let mut cfg = crate::Config::new(crate::PROTOCOL_VERSION).unwrap();
-        cfg.set_cc_algorithm(CongestionControlAlgorithm::CUBIC);
-
-        let mut r = Recovery::new(&cfg);
-
-        let mut now = Instant::now();
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-
-        // send out first packet (a full initcwnd).
-        let p = Sent {
-            pkt_num: 0,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 12000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 1);
-        assert_eq!(r.bytes_in_flight, 12000);
-
-        // First packet will be sent out immediately.
-        assert_eq!(r.congestion.pacer.rate(), 0);
-        assert_eq!(r.get_packet_send_time(), now);
-
-        // Wait 50ms for ACK.
-        now += Duration::from_millis(50);
-
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(0..1);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                10,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((0, 0, 12000))
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-        assert_eq!(r.bytes_in_flight, 0);
-        assert_eq!(r.rtt_stats.smoothed_rtt, Duration::from_millis(50));
-
-        // 1 MSS increased.
-        assert_eq!(r.cwnd(), 12000 + 1200);
-
-        // Send out second packet.
-        let p = Sent {
-            pkt_num: 1,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 6000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 1);
-        assert_eq!(r.bytes_in_flight, 6000);
-
-        // Pacing is not done during initial phase of connection.
-        assert_eq!(r.get_packet_send_time(), now);
-
-        // Send the third packet out.
-        let p = Sent {
-            pkt_num: 2,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 6000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 12000);
-
-        // Send the third packet out.
-        let p = Sent {
-            pkt_num: 3,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 3);
-        assert_eq!(r.bytes_in_flight, 13000);
-
-        // We pace this outgoing packet. as all conditions for pacing
-        // are passed.
-        let pacing_rate = (r.cwnd() as f64 * PACING_MULTIPLIER / 0.05) as u64;
-        assert_eq!(r.congestion.pacer.rate(), pacing_rate);
-
-        assert_eq!(
-            r.get_packet_send_time(),
-            now + Duration::from_secs_f64(12000.0 / pacing_rate as f64)
-        );
-    }
-
-    #[test]
-    fn pmtud_loss_on_timer() {
-        let mut cfg = crate::Config::new(crate::PROTOCOL_VERSION).unwrap();
-        cfg.set_cc_algorithm(CongestionControlAlgorithm::Reno);
-
-        let mut r = Recovery::new(&cfg);
-
-        let mut now = Instant::now();
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-
-        // Start by sending a few packets.
-        let p = Sent {
-            pkt_num: 0,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].in_flight_count, 1);
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 1);
-        assert_eq!(r.bytes_in_flight, 1000);
-
-        let p = Sent {
-            pkt_num: 1,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: true,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].in_flight_count, 2);
-
-        let p = Sent {
-            pkt_num: 2,
-            frames: smallvec![],
-            time_sent: now,
-            time_acked: None,
-            time_lost: None,
-            size: 1000,
-            ack_eliciting: true,
-            in_flight: true,
-            delivered: 0,
-            delivered_time: now,
-            first_sent_time: now,
-            is_app_limited: false,
-            tx_in_flight: 0,
-            lost: 0,
-            has_data: false,
-            pmtud: false,
-        };
-
-        r.on_packet_sent(
-            p,
-            packet::Epoch::Application,
-            HandshakeStatus::default(),
-            now,
-            "",
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].in_flight_count, 3);
-
-        // Wait for 10ms.
-        now += Duration::from_millis(10);
-
-        // Only the first  packets and the last one are acked.
-        let mut acked = ranges::RangeSet::default();
-        acked.insert(0..1);
-        acked.insert(2..3);
-
-        assert_eq!(
-            r.on_ack_received(
-                &acked,
-                25,
-                packet::Epoch::Application,
-                HandshakeStatus::default(),
-                now,
-                "",
-            ),
-            Ok((0, 0, 2 * 1000))
-        );
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.bytes_in_flight, 1000);
-        assert_eq!(r.congestion.lost_count, 0);
-
-        // Wait until loss detection timer expires.
-        now = r.loss_detection_timer().unwrap();
-
-        // Packet is declared lost.
-        r.on_loss_detection_timeout(HandshakeStatus::default(), now, "");
-        assert_eq!(r.epochs[packet::Epoch::Application].loss_probes, 0);
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 2);
-        assert_eq!(r.epochs[packet::Epoch::Application].in_flight_count, 0);
-        assert_eq!(r.bytes_in_flight, 0);
-        assert_eq!(r.cwnd(), 12000);
-
-        assert_eq!(r.congestion.lost_count, 0);
-
-        // Wait 1 RTT.
-        now += r.rtt();
-
-        r.detect_lost_packets(packet::Epoch::Application, now, "");
-
-        assert_eq!(r.epochs[packet::Epoch::Application].sent_packets.len(), 0);
-        assert_eq!(r.epochs[packet::Epoch::Application].in_flight_count, 0);
-        assert_eq!(r.bytes_in_flight, 0);
-        assert_eq!(r.congestion.lost_count, 0);
-    }
 }
